requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.7
lxml==5.2.2
MemePy==1.2.1
numpy==1.24.4
pandas==2.0.3
pillow==10.3.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
six==1.16.0
typing_extensions==4.11.0
tzdata==2024.1
Udacity-capstone-python==0.1
urllib3==2.2.1
